refactor: schema are restructured
